+++ hutool-core/src/main/java/cn/hutool/core/comparator/PinyinComparator.java
@@ -1,0 +1,29 @@
+package cn.hutool.core.comparator;
+
+import java.text.Collator;
+import java.util.Comparator;
+import java.util.Locale;
+
+/**
+ * 按照GBK拼音顺序对给定的汉字字符串排序
+ * 
+ * @author looly
+ * @since 4.0.8
+ */
+public class PinyinComparator implements Comparator<String> {
+
+	final Collator collator;
+
+	/**
+	 * 构造
+	 */
+	public PinyinComparator() {
+		collator = Collator.getInstance(Locale.CHINESE);
+	}
+
+	@Override
+	public int compare(String o1, String o2) {
+		return collator.compare(o1, o2);
+	}
+
+}